decomp patch: (basename ../../patches/decomp/017-basestadium-ulong-to-u32.patch .patch)

+++ src/Game/BasicStadium.cpp
@@ -619,7 +619,7 @@ void BasicStadium::LoadListOfGameplayInvisibleModels()
 {
     char szOptBin[80];
     nlSNPrintf(szOptBin, 0x50, "%s-GameplayInvisibleModels.bin", m_szBaseName);
-    unsigned long fileSize;
+    u32 fileSize;
     u32* pCur;
     u32* pData = (u32*)nlLoadEntireFile(szOptBin, &fileSize, 0x20, AllocateEnd);
     if (pData != NULL)
